rewrite commonTest functions to Cypress commands
